+++ src/napari_matplotlib/histogram.py
@@ -35,7 +35,6 @@ class HistogramWidget(SingleAxesWidget):
         Clear the axes and histogram the currently selected layer/slice.
         """
         layer = self.layers[0]
-        bins = np.linspace(np.min(layer.data), np.max(layer.data), 100)
 
         if layer.data.ndim - layer.rgb == 3:
             # 3D data, can be single channel or RGB
@@ -43,6 +42,12 @@ class HistogramWidget(SingleAxesWidget):
             self.axes.set_title(f"z={self.current_z}")
         else:
             data = layer.data
+        # Read data into memory if it's a dask array
+        data = np.asarray(data)
+
+        # Important to calculate bins after slicing 3D data, to avoid reading
+        # whole cube into memory.
+        bins = np.linspace(np.min(data), np.max(data), 100)
 
         if layer.rgb:
             # Histogram RGB channels independently