+++ src/napari_matplotlib/slice.py
@@ -1,18 +1,22 @@
-from typing import Any, Dict, Optional, Tuple
+from typing import Any, List, Optional, Tuple
 
 import matplotlib.ticker as mticker
 import napari
 import numpy as np
 import numpy.typing as npt
-from qtpy.QtWidgets import QComboBox, QHBoxLayout, QLabel, QSpinBox, QWidget
+from qtpy.QtCore import Qt
+from qtpy.QtWidgets import (
+    QComboBox,
+    QLabel,
+    QSlider,
+    QVBoxLayout,
+    QWidget,
+)
 
 from .base import SingleAxesWidget
 from .util import Interval
 
 __all__ = ["SliceWidget"]
-
-_dims_sel = ["x", "y"]
-_dims = ["x", "y", "z"]
 
 
 class SliceWidget(SingleAxesWidget):
@@ -31,27 +35,45 @@ class SliceWidget(SingleAxesWidget):
         # Setup figure/axes
         super().__init__(napari_viewer, parent=parent)
 
-        button_layout = QHBoxLayout()
-        self.layout().addLayout(button_layout)
-
         self.dim_selector = QComboBox()
+        self.dim_selector.addItems(["x", "y"])
+
+        self.slice_selector = QSlider(orientation=Qt.Orientation.Horizontal)
+
+        # Create widget layout
+        button_layout = QVBoxLayout()
         button_layout.addWidget(QLabel("Slice axis:"))
         button_layout.addWidget(self.dim_selector)
-        self.dim_selector.addItems(_dims)
-
-        self.slice_selectors = {}
-        for d in _dims_sel:
-            self.slice_selectors[d] = QSpinBox()
-            button_layout.addWidget(QLabel(f"{d}:"))
-            button_layout.addWidget(self.slice_selectors[d])
+        button_layout.addWidget(self.slice_selector)
+        self.layout().addLayout(button_layout)
 
         # Setup callbacks
-        # Re-draw when any of the combon/spin boxes are updated
+        # Re-draw when any of the combo/slider is updated
         self.dim_selector.currentTextChanged.connect(self._draw)
-        for d in _dims_sel:
-            self.slice_selectors[d].textChanged.connect(self._draw)
+        self.slice_selector.valueChanged.connect(self._draw)
 
         self._update_layers(None)
+
+    def on_update_layers(self) -> None:
+        """
+        Called when layer selection is updated.
+        """
+        if not len(self.layers):
+            return
+        if self.current_dim_name == "x":
+            max = self._layer.data.shape[-2]
+        elif self.current_dim_name == "y":
+            max = self._layer.data.shape[-1]
+        else:
+            raise RuntimeError("dim name must be x or y")
+        self.slice_selector.setRange(0, max - 1)
+
+    @property
+    def _slice_width(self) -> int:
+        """
+        Width of the slice being plotted.
+        """
+        return self._layer.data.shape[self.current_dim_index]
 
     @property
     def _layer(self) -> napari.layers.Layer:
@@ -61,7 +83,7 @@ class SliceWidget(SingleAxesWidget):
         return self.layers[0]
 
     @property
-    def current_dim(self) -> str:
+    def current_dim_name(self) -> str:
         """
         Currently selected slice dimension.
         """
@@ -74,36 +96,40 @@ class SliceWidget(SingleAxesWidget):
         """
         # Note the reversed list because in napari the z-axis is the first
         # numpy axis
-        return _dims[::-1].index(self.current_dim)
+        return self._dim_names.index(self.current_dim_name)
 
     @property
-    def _selector_values(self) -> Dict[str, int]:
+    def _dim_names(self) -> List[str]:
         """
-        Values of the slice selectors.
+        List of dimension names. This is a property as it varies depending on the
+        dimensionality of the currently selected data.
         """
-        return {d: self.slice_selectors[d].value() for d in _dims_sel}
+        if self._layer.data.ndim == 2:
+            return ["y", "x"]
+        elif self._layer.data.ndim == 3:
+            return ["z", "y", "x"]
+        else:
+            raise RuntimeError("Don't know how to handle ndim != 2 or 3")
 
     def _get_xy(self) -> Tuple[npt.NDArray[Any], npt.NDArray[Any]]:
         """
         Get data for plotting.
         """
-        x = np.arange(self._layer.data.shape[self.current_dim_index])
-
-        vals = self._selector_values
-        vals.update({"z": self.current_z})
+        val = self.slice_selector.value()
 
         slices = []
-        for d in _dims:
-            if d == self.current_dim:
+        for dim_name in self._dim_names:
+            if dim_name == self.current_dim_name:
                 # Select all data along this axis
                 slices.append(slice(None))
+            elif dim_name == "z":
+                # Only select the currently viewed z-index
+                slices.append(slice(self.current_z, self.current_z + 1))
             else:
                 # Select specific index
-                val = vals[d]
                 slices.append(slice(val, val + 1))
 
-        # Reverse since z is the first axis in napari
-        slices = slices[::-1]
+        x = np.arange(self._slice_width)
         y = self._layer.data[tuple(slices)].ravel()
 
         return x, y
@@ -115,7 +141,7 @@ class SliceWidget(SingleAxesWidget):
         x, y = self._get_xy()
 
         self.axes.plot(x, y)
-        self.axes.set_xlabel(self.current_dim)
+        self.axes.set_xlabel(self.current_dim_name)
         self.axes.set_title(self._layer.name)
         # Make sure all ticks lie on integer values
         self.axes.xaxis.set_major_locator(