+++ src/napari_matplotlib/tests/test_theme.py
@@ -1,10 +1,15 @@
+import os
+import shutil
 from copy import deepcopy
+from pathlib import Path
 
+import matplotlib
 import napari
 import numpy as np
 import pytest
+from matplotlib.colors import to_rgba
 
-from napari_matplotlib import ScatterWidget
+from napari_matplotlib import HistogramWidget, ScatterWidget
 from napari_matplotlib.base import NapariMPLWidget
 
 
@@ -32,9 +37,11 @@ def _mock_up_theme() -> None:
     https://napari.org/stable/gallery/new_theme.html
     """
     blue_theme = napari.utils.theme.get_theme("dark", False)
-    blue_theme.name = "blue"
+    blue_theme.label = "blue"
     blue_theme.background = "#4169e1"  # my favourite shade of blue
-    napari.utils.theme.register_theme("blue", blue_theme)
+    napari.utils.theme.register_theme(
+        "blue", blue_theme, source="napari-mpl-tests"
+    )
 
 
 def test_theme_background_check(make_napari_viewer):
@@ -140,3 +147,46 @@ def test_custom_theme(make_napari_viewer, theme_path, brain_data):
     viewer.layers.selection.add(viewer.layers[1])
 
     return deepcopy(widget.figure)
+
+
+def find_mpl_stylesheet(name: str) -> Path:
+    """Find the built-in matplotlib stylesheet."""
+    return Path(matplotlib.__path__[0]) / f"mpl-data/stylelib/{name}.mplstyle"
+
+
+def test_custom_stylesheet(make_napari_viewer, image_data):
+    """
+    Test that a stylesheet in the current directory is given precidence.
+
+    Do this by copying over a stylesheet from matplotlib's built in styles,
+    naming it correctly, and checking the colours are as expected.
+    """
+    # Copy Solarize_Light2 as if it was a user-overriden stylesheet.
+    style_sheet_path = (
+        Path(matplotlib.get_configdir()) / "napari-matplotlib.mplstyle"
+    )
+    if style_sheet_path.exists():
+        pytest.skip("Won't ovewrite existing custom style sheet.")
+    shutil.copy(
+        find_mpl_stylesheet("Solarize_Light2"),
+        style_sheet_path,
+    )
+
+    try:
+        viewer = make_napari_viewer()
+        viewer.add_image(image_data[0], **image_data[1])
+        widget = HistogramWidget(viewer)
+        assert widget.mpl_style_sheet_path == style_sheet_path
+        ax = widget.figure.gca()
+
+        # The axes should have a light brownish grey background:
+        assert ax.get_facecolor() == to_rgba("#eee8d5")
+        assert ax.patch.get_facecolor() == to_rgba("#eee8d5")
+
+        # The figure background and axis gridlines are light yellow:
+        assert widget.figure.patch.get_facecolor() == to_rgba("#fdf6e3")
+        for gridline in ax.get_xgridlines() + ax.get_ygridlines():
+            assert gridline.get_visible() is True
+            assert gridline.get_color() == "#fdf6e3"
+    finally:
+        os.remove(style_sheet_path)