+++ docs/changelog.rst
@@ -1,7 +1,40 @@
 Changelog
 =========
+1.0.3
+-----
+Changes
+~~~~~~~
+- The slice widget is now limited to slicing along the x/y dimensions. Support
+  for slicing along z has been removed for now to make the code simpler.
+- The slice widget now uses a slider to select the slice value.
 
-0.3.0
+Bug fixes
+~~~~~~~~~
+- Fixed creating 1D slices of 2D images.
+- Removed the limitation that only the first 99 indices could be sliced using
+  the slice widget.
+
+1.0.2
+-----
+Bug fixes
+~~~~~~~~~
+- A full dataset is no longer read into memory when using ``HistogramWidget``.
+  Only the current slice is loaded.
+- Fixed compatibility with napari 0.4.18.
+
+Changes
+~~~~~~~
+- Histogram bin limits are now caclualted from the slice being histogrammed, and
+  not the whole dataset. This is as a result of the above bug fix.
+
+1.0.1
+-----
+Bug fixes
+~~~~~~~~~
+- Pinned that maximum version of `napari` to 0.4.17, since ``napari-matplotlib``
+  does not yet work with ``napari`` 0.4.18.
+
+1.0.0
 -----
 
 New features
@@ -15,6 +48,8 @@ Visual improvements
 ~~~~~~~~~~~~~~~~~~~
 - The background of ``napari-matplotlib`` figures and axes is now transparent, and the text and axis colour respects the ``napari`` theme.
 - The icons in the Matplotlib toolbar are now the same size as icons in the napari window.
+- Custom style sheets can now be set to customise plots. See the user guide
+  for more information.
 
 Changes
 ~~~~~~~